requirements.txt
Flask
numpy
scikit-learn
pandas
mysql-connector-python

